requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
uwsgi
mysqlclient
rejson
